+++ src/components/detail/post/application/Calandar.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Calandar = () => {
+  return <div></div>;
+};
+
+export default Calandar;
