+++ src/components/detail/post/application/TimeCard.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const TimeCard = () => {
+  return <div></div>;
+};
+
+export default TimeCard;
